Workaround: use old pthread mutex code
The new mutex code can't be used as-is with the old kernel psynch code; it has a tendency to skip wakeups when used like that. To use it, we'd have to either modify the new mutex code (not what we want), add some hack in the old kernel psynch code (also not what we want), or use the new kernel psynch code.

That last option is really the most preferable one, but the problem is that it requires some fiddling around with the LKM and turnstiles, which I don't want to get into right now (especially given that the LKM is mostly stable right now).

So instead, the workaround is to just keep using the old mutex code.